+++ src/app/pages/pass-finalisation/cat-adi-part3/components/change-start-end-time-modal/change-start-end-time-modal.ts
@@ -1,0 +1,25 @@
+import { Component } from '@angular/core';
+import {ModalController} from '@ionic/angular';
+
+@Component({
+  selector: 'terminate-test-modal',
+  templateUrl: './change-start-end-time-modal.html',
+  styleUrls: ['./change-start-end-time-modal.scss'],
+})
+export class ChangeStartEndTimeModal {
+  startTime = '';
+  endTime = '';
+
+  constructor(
+    private modalController: ModalController,
+  ) {
+  }
+
+  async onCancel() {
+    await this.modalController.dismiss()
+  }
+
+  async onConfirm() {
+    await this.modalController.dismiss({startTime: this.startTime, endTime: this.endTime})
+  }
+}
