+++ src/app/pages/pass-finalisation/cat-adi-part3/components/pass-finalisation.cat-adi3.components.module.ts
@@ -7,9 +7,12 @@ import { IonicModule } from '@ionic/angular';
 import { FurtherDevelopmentComponent } from '@pages/pass-finalisation/cat-adi-part3/components/further-development/further-development.component';
 import { ReasonGivenComponent } from '@pages/pass-finalisation/cat-adi-part3/components/reason-given/reason-given.component';
 import { TestStartEndTimesComponent } from '@pages/pass-finalisation/cat-adi-part3/components/test-start-end-times/test-start-end-times';
+import {
+  ChangeStartEndTimeModal
+} from '@pages/pass-finalisation/cat-adi-part3/components/change-start-end-time-modal/change-start-end-time-modal';
 
 @NgModule({
-  declarations: [FurtherDevelopmentComponent, ReasonGivenComponent, TestStartEndTimesComponent],
+  declarations: [FurtherDevelopmentComponent, ReasonGivenComponent, TestStartEndTimesComponent, ChangeStartEndTimeModal],
   imports: [IonicModule, CommonModule, ComponentsModule, ReactiveFormsModule, DirectivesModule],
   exports: [FurtherDevelopmentComponent, ReasonGivenComponent, TestStartEndTimesComponent],
 })
