+++ src/app/pages/pass-finalisation/cat-adi-part3/components/test-start-end-times/test-start-end-times.html
@@ -51,7 +51,7 @@
       <ion-text class="mes-data">Approx. duration:</ion-text>
     </ion-row>
   </ion-col>
-  <ion-col class="centre-col">
+  <ion-col class="direction-column centre-col">
     <ion-row>
       <ion-text class="mes-data">{{formatTime(startTime)}}</ion-text>
     </ion-row>
