+++ src/app/pages/pass-finalisation/cat-adi-part3/components/test-start-end-times/test-start-end-times.scss
@@ -8,3 +8,7 @@
   align-items: center;
   justify-content: center;
 }
+
+.direction-column {
+  flex-direction: column;
+}
