--- src/components/common/change-start-end-time-modal/change-start-end-time-modal.module.ts
@@ -1,11 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { IonicModule } from '@ionic/angular';
-
-import { TerminateTestModal } from '@components/common/terminate-test-modal/terminate-test-modal';
-
-@NgModule({
-  imports: [CommonModule, IonicModule],
-  declarations: [TerminateTestModal],
-})
-export class ChangeStartEndTimeModalModule {}
